First split-up of wifite.py
